+++ lib/presentation/views/views.dart
@@ -1,0 +1,3 @@
+export 'package:cinemapedia/presentation/views/movies/favorites_view.dart';
+export 'package:cinemapedia/presentation/views/movies/home_view.dart';
+
